Added should update checker on move files.

+++ src/com/severn/files/GcsMove.groovy
@@ -1,23 +1,32 @@
 import java.util.logging.*
-import com.google.appengine.tools.cloudstorage.*
 
-def FROM_BUCKET = 'wild-ride-app-viber-temp-restored-events'
-def TO_BUCKET   = 'wild-ride-viber-app-events-success-archive-1'
-def TABLE       = 'raw_sessions_end' + '/'// + '/2016/01/1'
+import com.google.appengine.tools.cloudstorage.*
+import com.severn.script.utils.GcsUtils;
+
+def FROM_BUCKET = 'wild-ride-events-fail-archive-1'
+def TO_BUCKET   = 'wild-ride-stage-1-temp'
+def TABLE       = 'push_notification' + '/' + '2016/0'
 def count = 0
 Logger logger = Logger.getLogger('com.severn')
 
-GcsService service = GcsServiceFactory.createGcsService();
+GcsService service = GcsUtils.getRetryGcsService()
 ListOptions options = new ListOptions.Builder().setPrefix(TABLE).setRecursive(true).build();
 ListResult files = service.list(FROM_BUCKET, options);
 
-files.each {
-    GcsFilename initialFile = new GcsFilename(FROM_BUCKET, it.getName());
-    GcsFilename newFile = new GcsFilename(TO_BUCKET, it.getName());
-    logger.fine("About ot copy from $initialFile to $newFile")
-    service.copy(initialFile, newFile);
-    service.delete(initialFile);
-    count++
+def willCopy = { ListItem it ->
+    def name = it.getName()
+    name.endsWith('.restored')
+}
+
+files.each { ListItem it ->
+    if (willCopy(it)) {
+        GcsFilename initialFile = new GcsFilename(FROM_BUCKET, it.getName());
+        GcsFilename newFile = new GcsFilename(TO_BUCKET, it.getName());
+        logger.fine("About ot copy from $initialFile to $newFile")
+        service.copy(initialFile, newFile);
+        service.delete(initialFile);
+        count++
+    }
 }
 
 'OK : ' + count
